434 - use unsigned short return type everywhere in render methods of controllers

+++ includes/EditorMenuController.hpp
@@ -73,11 +73,9 @@ public:
      *
      * @param context commons items for controller
      *
-     * @return uint8_t   id of the next screen controller
+     * @return unsigned short id of the next screen controller
      */
-    uint8_t render(
-        utils::Context& context
-    );
+    unsigned short render(utils::Context& context);
 
     /**
      * @brief change the position of the cursor in the menu
